pyproject.toml
[tool.poetry]
name = "servcy-server"
version = "0.1.0"
description = ""
authors = ["Megham Garg <megham@servcy.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.1"
django = "^4.2"
django-cors-headers = "^3.14.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
boto3 = "^1.26.118"
django-crontab = "^0.7.1"
django-storages = "^1.13.2"
pydantic = "^1.10.7"
mypy = "^1.2.0"
psycopg2 = "^2.9.6"
configparser = "^5.3.0"
django-crum = "^0.7.9"
twilio = "^8.1.0"
django-filter = "^23.1"
gunicorn = "^20.1.0"
serpy = "^0.3.1"
cryptography = "^41.0.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
coverage = "^7.2.3"
pre-commit = "^3.2.2"
docformatter = "^1.6.0"
docstr-coverage = "^2.3.0"
pylint = "^2.17.2"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
